compose_flask/requirements.txt
flask
mysql-connector
redis